modificaciones

+++ src/main/java/me/gonzager/ex/operaciones/drones/DronSeguridad.java
@@ -4,11 +4,11 @@ import me.gonzager.ex.operaciones.misiones.Mision;
 
 public class DronSeguridad extends Dron  {
 
-    public DronSeguridad(Double autonomia, Integer procesamiento) {
+    public DronSeguridad(Integer autonomia, Integer procesamiento) {
         super(autonomia, procesamiento);
     }
 
-    public DronSeguridad(Double autonomia, Integer procesamiento, Mision mision) {
+    public DronSeguridad(Integer autonomia, Integer procesamiento, Mision mision) {
         super(autonomia, procesamiento, mision);
     }
